Migrate timer into shared service.
Reduces most extensions by about 1KB.
