+++ tools/experiments/experiments.js
@@ -22,6 +22,8 @@ import {getMode} from '../../src/mode';
 import {isExperimentOn, toggleExperiment} from '../../src/experiments';
 import {listenOnce} from '../../src/event-helper';
 import {onDocumentReady} from '../../src/document-ready';
+//TODO(@cramforce): For type. Replace with forward declaration.
+import '../../src/service/timer-impl';
 
 const COOKIE_MAX_AGE_DAYS = 180;  // 6 month
